Explicitly request ITK packages
At the moment if ITK is built against VTK (primarily for VTKGlue), if a project uses ITK but not VTK there are linker errors. A patch was submitted this morning to ITK to resolve this, but it may take a while to agree on the implementation details. Once the patch is merged, if an ITK build later than that patch is used, the FIND_PACKAGE line can be changed back to simply (ITK REQUIRED).